moving kde to submodule
