+++ chaospy/distributions/kernel/__init__.py
@@ -1,0 +1,2 @@
+"""Kernel density estimation."""
+from .gaussian import GaussianKDE
